feat: add thermometer in test and RPi mode
