+++ smart_lamp/server.json
@@ -1,1 +1,1 @@
-{"data-source-ids": {"state": "2238658e-b4a3-4ef7-9b48-4406db3ac1b3"}, "id": "3711067a-050e-4bb0-aa84-8475baef2c02", "port": "8000", "url": "127.0.0.1"}
+{"data-source-ids": {"state": "8d958c97-34b0-4e85-8995-2cca83bba873"}, "id": "f04de171-9539-4c4a-9922-b67d0b392990", "port": "8000", "url": "127.0.0.1"}
