+++ temperature_sensor/config.json
