+++ temperature_sensor/server.json
